backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pandas==2.2.3
scikit-learn==1.5.2
fairlearn==0.10.0
shap==0.46.0
anthropic==0.36.2
reportlab==4.2.5
aiofiles==24.1.0
httpx==0.27.2
numpy==1.26.4
scipy==1.14.1
matplotlib==3.9.2
sentence-transformers==3.3.1
